Adds undo buttons back. Adjusts css
