+++ frontend/src/js/components/PlotTools/PlotTools.scss
@@ -1,0 +1,6 @@
+.material-icons-button{
+    padding: 3px 7px 2px 7px;
+    .material-icons{
+        font-size: 18px;
+    }
+}
