+++ frontend/src/js/containers/RightSideBar.jsx
@@ -1,5 +1,5 @@
 import React, {Component} from 'react'
-import PlotTools from '../components/PlotTools.jsx'
+import PlotTools from '../components/PlotTools/PlotTools.jsx'
 /* global $ */
 
 class RightSideBar extends Component{
@@ -7,10 +7,17 @@ class RightSideBar extends Component{
     render(){
         return (
             <div id='right-side-bar'>
-                <PlotTools />
+                <PlotTools {...this.props}/>
             </div>
         )
     }
+}
+
+RightSideBar.propTypes = {
+    onUndo: React.PropTypes.func,
+    onRedo: React.PropTypes.func,
+    undoEnabled: React.PropTypes.bool,
+    redoEnabled: React.PropTypes.bool,
 }
 
 export default RightSideBar
